+++ apps/admin-ui-backup/src/views/backup/configuration/backup-configuration.tsx
@@ -11,6 +11,7 @@ import {
   LabeledValue,
   ListRow,
   Padding,
+  RouteLeavingGuard,
   Row,
   Select,
   Switch,
@@ -56,8 +57,6 @@ import {
   ZIMBRA_ADMIN_URN,
 } from '../../../constants';
 import { fetchSoap } from '../../../services/bucket-service';
-import { useBackupStore } from '../../../store/backup/store';
-import { RouteLeavingGuard } from '../../ui-extras/nav-guard';
 
 const BackupConfiguration: FC = () => {
   const { server } = useParams();
@@ -106,7 +105,6 @@ const BackupConfiguration: FC = () => {
   const [manageExternalVolumeNewLocalMountpoint, setManageExternalVolumeNewLocalMountpoint] =
     useState<string>('');
   const [rootVolumePath, setRootVolumePath] = useState<string>('');
-  const selectedBackupServer = useBackupStore((state) => state.selectedServer);
   const { data: rights } = useCurrentUserRights();
   const allowSetBackup = useMemo(() => {
     const rightsConfig = find(rights, { type: CONFIG }) || { all: [], type: CONFIG };
@@ -966,7 +964,7 @@ const BackupConfiguration: FC = () => {
                   crossAlignment="flex-start"
                 >
                   <ds-text as="h2" size="medium" weight="bold" color="gray0">
-                    {selectedBackupServer}{' '}
+                    {server}{' '}
                     {t('backup.backup_configuration', 'backup configuration')}
                   </ds-text>
                 </Row>
@@ -1589,15 +1587,7 @@ const BackupConfiguration: FC = () => {
             </ListRow>
           </Container>
         </Container>
-        <RouteLeavingGuard when={isDirty} onSave={onSave}>
-          <ds-text as="p">
-            {t(
-              'label.unsaved_changes_line1',
-              'Are you sure you want to leave this page without saving?',
-            )}
-          </ds-text>
-          <ds-text as="p">{t('label.unsaved_changes_line2', 'All your unsaved changes will be lost')}</ds-text>
-        </RouteLeavingGuard>
+        <RouteLeavingGuard when={isDirty} onSave={onSave} />
       </Container>
     </>
   );