+++ apps/admin-ui-domains/src/views/domain/details/domain-2fa.tsx
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import { 	Button,	Container,	Padding,	Row,	useSnackbar } from '@zextras/ui-components';
+import { 	Button,	Container,	Padding,	RouteLeavingGuard,	Row,	useSnackbar } from '@zextras/ui-components';
 import {  differenceWith, isEqual, map, some  } from 'lodash-es';
 import {  FC, useCallback, useEffect, useMemo, useState  } from 'react';
 import {  useTranslation  } from 'react-i18next';
@@ -13,7 +13,6 @@ import {  OK  } from '../../../constants';
 import { useSelectedDomain } from '../../../hooks/use-selected-domain';
 import {  list2faPolicies  } from '../../../services/list-2fa-policies';
 import {  set2faPolicies  } from '../../../services/set-2fa-policies';
-import {  RouteLeavingGuard  } from '../../ui-extras/nav-guard';
 import {  isValidIpRange,TwoFactorPolicyArray  } from '../../utility/utils';
 import {  TwoFactorAuthencationConfig  } from '../two-factor-authentication/2fa-config';
 
@@ -197,15 +196,7 @@ const DomainTwoFactorAuthentication: FC = () => {
 					twoFactorPolicyArray={twoFactorPolicyArray}
 				/>
 			</Container>
-			<RouteLeavingGuard when={isDirty} onSave={handleOnSave}>
-				<ds-text as="p">
-					{t(
-						'label.unsaved_changes_line1',
-						'Are you sure you want to leave this page without saving?'
-					)}
-				</ds-text>
-				<ds-text as="p">{t('label.unsaved_changes_line2', 'All your unsaved changes will be lost')}</ds-text>
-			</RouteLeavingGuard>
+			<RouteLeavingGuard when={isDirty} onSave={handleOnSave} />
 		</Container>
 	);
 };